+++ CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Changed
+- Do not prefill feed on empty subscriptions
+
 ## [0.0.7] - 2023-10-10
 
 ### Added